[#38] refactor: common.utils 패키지 생성 및 JwtUtils를 이동
- 프로젝트 전체에서 사용하므로 common에 있는게 맞다고 판단
